+++ WebApplication1/Controllers/HomeController.cs
@@ -28,41 +28,6 @@ namespace WebApplication1.Controllers
             return View(resources);
         }
 
-
-        [HttpGet("{id:length(24)}", Name = "GetResource")]
-        public ActionResult<Resource> Get(string id)
-        {
-            var resource = _resourceService.Get(id);
-
-            if (resource == null)
-            {
-                return NotFound();
-            }
-
-            return resource;
-        }
-
-        // POST: Resource/Create
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public ActionResult<Resource> Create(Resource resource)
-        {
-            try
-            {
-                if (resource == null){
-                    resource = new Resource { ResourceTitle = "Test Overflow", ResourceDescription = "This is Create test", Uri = "www.testoverflow.com", UpVote = 11, DownVote = 10 };
-                }
-                // TODO: Add insert logic here
-                _resourceService.Create(resource);
-
-                return CreatedAtRoute("GetResource", new { id = resource.Id.ToString() }, resource);
-            }
-            catch
-            {
-                return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-            }
-        }
-
         public IActionResult Privacy()
         {
             return View();
