+++ WebApplication1/Startup.cs
@@ -70,6 +70,7 @@ namespace WebApplication1
 
             app.UseEndpoints(endpoints =>
             {
+                //endpoints.MapControllers();
                 endpoints.MapDefaultControllerRoute();
                 // replaces the following code
