+++ WebApplication1/Views/Shared/_Layout.cshtml
@@ -85,11 +85,11 @@
                 </div>
                 <div class="row my-5">
                     <h1 class="display-4 mr-auto">Your Team's Resources</h1>
-                    <form class="form-inline my-2 my-lg-0" action="/get", method="get">
+                    <form class="form-inline my-2 my-lg-0" action="/resource", method="get">
                         <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search" id="id" name="id">
                         <button class="btn btn-outline-info my-2 my-sm-0" type="submit">Search</button>
                     </form>
-                    <form class="form-inline my-2 my-lg-0" action="/create" method="post">
+                    <form class="form-inline my-2 my-lg-0" action="/resource/create" method="post">
                         <button class="btn btn-info form-inline mx-2 font-weight-bold" type="submit">Add resource</button>
                     </form>
                 </div>
